refactoring into BaseEvaluationResults class
